+++ winter/src/test/kotlin/io/jentz/winter/GraphTest.kt
@@ -7,7 +7,6 @@ import io.jentz.winter.plugin.SimplePlugin
 import io.kotlintest.matchers.boolean.shouldBeFalse
 import io.kotlintest.matchers.boolean.shouldBeTrue
 import io.kotlintest.matchers.collections.shouldContainAll
-import io.kotlintest.matchers.collections.shouldHaveSize
 import io.kotlintest.matchers.types.shouldBeInstanceOf
 import io.kotlintest.matchers.types.shouldBeNull
 import io.kotlintest.matchers.types.shouldBeSameInstanceAs
@@ -545,42 +544,55 @@ class GraphTest {
     }
 
     @Nested
-    @DisplayName("#*OfType methods")
+    @DisplayName("#*OfType")
     inner class OfTypeMethods {
 
-        private val testComponent = component {
+        private val testGraph = graph {
             prototype("something else") { Any() }
             prototype("a") { "a" }
             prototype("b") { "b" }
             prototype("c") { "c" }
+            prototype { "bar" }
+
+            setOfType<String>("a")
+            setOfProvidersForType<String>("b")
+            mapOfType<String>("c", defaultKey = "foo")
+            mapOfProvidersForType<String>("d", defaultKey = "foo")
         }
 
         @Test
-        fun `#providersOfTypeByKey should fail if type key is not correct`() {
-            shouldThrow<IllegalArgumentException> {
-                testComponent.createGraph().providersOfTypeByKey(typeKey<String>())
-            }
+        fun `should provide a set of instances of type`() {
+            val set: Set<String> = testGraph.instance("a", true)
+            set.shouldBe(setOf("a", "b", "c", "bar"))
         }
 
         @Test
-        fun `#instancesOfTypeByKey should fail if type key is not correct`() {
-            shouldThrow<IllegalArgumentException> {
-                testComponent.createGraph().instancesOfTypeByKey(typeKey<String>())
-            }
+        fun `should provide a set of providers of type`() {
+            val set: Set<Provider<String>> = testGraph.instance("b", true)
+            set.map { it() }.shouldContainAll("a", "b", "c", "bar")
         }
 
         @Test
-        fun `#providersOfType should return a set of providers of a given type`() {
-            val providers = testComponent.createGraph().providersOfType<String>()
-            providers.shouldHaveSize(3)
-            providers.map { it() }.shouldContainAll("a", "b", "c")
+        fun `should provide a map of instances of type`() {
+            val map: Map<Any, String> = testGraph.instance("c", true)
+            map.shouldBe(mapOf(
+                "a" to "a",
+                "b" to "b",
+                "c" to "c",
+                "foo" to "bar"
+            ))
         }
 
         @Test
-        fun `#instancesOfType should return a set of instances of given type`() {
-            val instances = testComponent.createGraph().instancesOfType<String>()
-            instances.shouldHaveSize(3)
-            instances.shouldContainAll("a", "b", "c")
+        fun `should provide a map of providers of type`() {
+            val map: Map<Any, Provider<String>> = testGraph.instance("d", true)
+            map.map { (k, v) -> k to v() }.toMap()
+                .shouldBe(mapOf(
+                    "a" to "a",
+                    "b" to "b",
+                    "c" to "c",
+                    "foo" to "bar"
+                ))
         }
 
     }
@@ -653,7 +665,7 @@ class GraphTest {
 
         @Test
         fun `should initialize graph with given component`() {
-            Graph(WinterApplication(),null, emptyComponent, null, null)
+            Graph(WinterApplication(), null, emptyComponent, null, null)
                 .component.shouldBe(emptyComponent)
         }
 
@@ -673,7 +685,7 @@ class GraphTest {
 
         @Test
         fun `should derive component when builder block is given`() {
-            val graph = Graph(Winter,null, emptyComponent, null) { constant(42) }
+            val graph = Graph(Winter, null, emptyComponent, null) { constant(42) }
             graph.instance<Int>().shouldBe(42)
         }